+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -62,6 +62,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Removed
 * `objc2-metal`: Removed internal `__MTLPackedFloat3` and made `MTLPackedFloat3` public.
+* `objc2-core-data`: Removed broken GNUStep support.
+* `objc2-quartz-core`: Removed broken GNUStep support.
 
 
 ## 0.2.0 - 2024-04-17